+++ src/widgets/helper/ChannelView.cpp
@@ -125,6 +125,10 @@ namespace {
         {
             addPageLink("Homies");
         }
+        else if (creatorFlags.has(MessageElementFlag::TwitchEmote))
+        {
+            addPageLink("Twitch");
+        }
     }
 }  // namespace